Renames EtherscanButton to ScanBlockButton
